+++ python/oneflow/test/graph/test_free_tensor_not_in_job.py
@@ -1,0 +1,62 @@
+"""
+Copyright 2020 The OneFlow Authors. All rights reserved.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+"""
+import unittest
+import numpy as np
+
+import oneflow as flow
+import oneflow.unittest
+import oneflow.nn as nn
+
+
+class my_model(nn.Module):
+    def __init__(self):
+        super(my_model, self).__init__()
+        self.batch_norm = nn.BatchNorm1d(6)
+
+    def forward(self, x):
+        y = self.batch_norm(x)
+        return y
+
+
+model = my_model()
+model.eval()
+model.to_global(flow.env.all_device_placement("cpu"), flow.sbp.broadcast)
+
+
+class Testgraph(flow.nn.Graph):
+    def __init__(self, model):
+        super(Testgraph, self).__init__()
+        self.module = model
+
+    def build(self, x):
+        return self.module(x)
+
+
+test_graph = Testgraph(model)
+
+
+@flow.unittest.skip_unless_1n1d()
+class TestFreeTensorNotInJob(flow.unittest.TestCase):
+    def test_free_tensor_not_in_job(test_case):
+        x = flow.randn(1, 6, 2).to_global(
+            placement=flow.env.all_device_placement("cpu"), sbp=flow.sbp.split(0)
+        )
+        y = test_graph(x)
+        test_case.assertEqual(y.size(), (1, 6, 2))
+
+
+if __name__ == "__main__":
+    unittest.main()
